+++ src/p_view.c
@@ -1707,7 +1707,7 @@ void ClientEndServerFrame (edict_t *ent)
 	for (i=0 ; i<3 ; i++)
 	{
 		current_client->ps.pmove.origin[i] = ent->s.origin[i]*8.0;
-		current_client->ps.pmove.velocity[i] = ent->velocity[i]*8.0;
+		current_client->ps.pmove.velocity[i] = (int)(ent->velocity[i]*8.0);
 	}
 
 	//
